Do not store multistart options in separate module
