+++ src/optimagic/__init__.py
@@ -8,11 +8,11 @@ from optimagic.benchmarking.get_benchmark_problems import get_benchmark_problems
 from optimagic.benchmarking.run_benchmark import run_benchmark
 from optimagic.differentiation.derivatives import first_derivative, second_derivative
 from optimagic.logging.read_log import OptimizeLogReader
+from optimagic.optimization.multistart import MultistartOptions
 from optimagic.optimization.optimize import maximize, minimize
 from optimagic.optimization.optimize_result import OptimizeResult
 from optimagic.parameters.bounds import Bounds
 from optimagic.parameters.constraint_tools import check_constraints, count_free_params
-from optimagic.parameters.multistart import MultistartOptions
 from optimagic.parameters.scaling import ScalingOptions
 from optimagic.visualization.convergence_plot import convergence_plot
 from optimagic.visualization.history_plots import criterion_plot, params_plot
